Moving progress bar classes to a new place
