+++ src/Console/ProgressBar/ProgressBarFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace BumbleDocGen\Core\Console;
+namespace BumbleDocGen\Console\ProgressBar;
 
 use Symfony\Component\Console\Style\OutputStyle;
